+++ cartographer/mapping_2d/sparse_pose_graph/constraint_builder.cc
@@ -43,7 +43,7 @@ transform::Rigid2d ComputeSubmapPose(const Submap& submap) {
 }
 
 ConstraintBuilder::ConstraintBuilder(
-    const mapping::sparse_pose_graph::proto::ConstraintBuilderOptions& options,
+    const mapping::pose_graph::proto::ConstraintBuilderOptions& options,
     common::ThreadPool* const thread_pool)
     : options_(options),
       thread_pool_(thread_pool),
